(MODULES-6697) Implement packageprovider type
This commit adds a new type and provider for managing the PowerShellGet
package providers. **Note** The official API at this time does not allow
removing PowerShellGet PackageProviders. There is a way to delete the
source_location to remove a PackageProvider, but this is an unsupported
method that may cause side effects. Thus this provider only implements
create and update.