Allow SecurityProtocol to be specified for a repo
A parameter (securityprotocols) has been added to the psrepository type
which will accept a list of security protocols that the repository will
accept, the list can contain any values can be specified for the
ServicePointManager.SecurityProtocol Property (see
https://docs.microsoft.com/en-us/dotnet/api/system.net.servicepointmanager.securityprotocol)

e.g.

    psrepository { 'psrepo':
        ensure              => present,
        source_location     => 'https://local.repo.domain/',
        installation_policy => 'trusted',
        securityprotocols   => [ TLS11, TLS12 ]
    }

This is to resolve an SSL connection issue that can occur if the
PowerShell repository enforces the use of a specific TLS version, but
the Windows client has a default protocol version that is lower.